requirements.txt
requests==2.32.5
httpx==0.28.1
pandas==2.3.1
python-dotenv==1.1.1
PyYAML==6.0.2
lxml==6.0.0
typing-extensions==4.14.1
yt-dlp==2025.8.11
dashscope==1.24.1
datetime==5.5
langchain-openai==0.3.31
langchain-core==0.3.74
langsmith==0.4.16
langgraph==0.6.6
pydantic==2.11.7
pydantic-core==2.33.2
lark-oapi==1.4.21
deepagents==0.0.4
aiohttp==3.12.15
weasyprint==66.0
reportlab==4.4.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
pygments==2.19.2
langchain-mcp-adapters==0.1.9
fastmcp==2.11.3
openpyxl==3.1.5